+++ src/Data/Artifacts.cs
@@ -1,0 +1,45 @@
+
+using System.Collections.Concurrent;
+
+namespace BetterMagnets.Data;
+
+public static class Artifacts {
+    public static readonly ConcurrentDictionary<int, ArtifactSkillData> DataList = new();
+
+    // public const int ArtifactSkill_BeefyBoy = 39278;
+    // public const int ArtifactSkill_DivingForDummies = 30350;
+    // public const int ArtifactSkill_OneShotOneKill = 33914;
+    // public const int ArtifactSkill_ShootFastGoSlow = 38424;
+    // public const int BookOfExperienceArtifactData = 32448;
+    // public const int ChemistKitArtifactData = 44866;
+    public const int CompanyIssuedMagnetArtifactData = 39198;
+    // public const int CurrencyFromBlockArtifactData_Gold = 46132;
+    // public const int CurrencyFromBlockArtifactData_Nitra = 46492;
+    // public const int CurrencyFromBlockArtifactData_XP = 43512;
+    // public const int CurrencyOnDamageTakenArtifactData_XP = 30274;
+    // public const int DamageShieldArtifactData_Cryo = 32140;
+    // public const int DamageShieldArtifactData_Fire = 37250;
+    // public const int ExtendMiningAreaData = 32038;
+    // public const int HealOnDropPodEnterArtifactData = 44460;
+    // public const int HealOnLevelUpArtifactData = 37816;
+    // public const int JetBootsArtifactData = 32148;
+    // public const int LuckRerollArtifactData = 46518;
+    // public const int MoveStateBuffArtifactData_Dodger = 41586;
+    // public const int MoveStateBuffArtifactData_Spraying = 42586;
+    // public const int MoveStateBuffArtifactData_Sprinter = 40268;
+    // public const int MultiToolArtifactData = 38870;
+    // public const int OverOverclockArtifactData = 44538;
+    // public const int Pay2WinArtifactData = 38504;
+    // public const int PiercingProjectilesArtifactData = 43326;
+    // public const int RedSugarPickStatModArtifactData_MaxHp = 39690;
+    // public const int ScaleStatWithMissingHpArtifactData_Armor = 44546;
+    // public const int ScaleStatWithMissingHpArtifactData_Dmg = 40946;
+    // public const int ScaleStatWithWalletArtifactData_GoldToDmg = 40690;
+    // public const int ScaleStatWithWalletArtifactData_NitraToCritChance = 40624;
+    // public const int ScaleStatWithWalletArtifactData_NitraToDmgMoveSpeed = 43840;
+    // public const int ShopDiscountArtifactData = 33618;
+    // public const int SpawnHuuliArtifactData = 42268;
+    // public const int StatModPrPlayerLevelArtifactData = 39726;
+    // public const int TimedBuffOnDamageTakenArtifactData_Dodge = 45114;
+    // public const int WeaponBoxArtifactData = 46090;
+}