+++ src/Data/Controllers.cs
@@ -1,0 +1,7 @@
+
+
+namespace BetterMagnets.Data;
+
+public static class Controllers {
+    public static GameController GameController { get; set; }
+}