+++ src/Settings/Config.cs
@@ -6,7 +6,7 @@ namespace BetterMagnets.Settings;
 public class Config {
     public static void Load(ConfigFile configFile, ManualLogSource logger, string worldType) {
         // Settings setup
-        ENV.Settings.Setup();
+        ENV.General.Setup();
         Utils.Settings.Config.Setup(MyPluginInfo.PLUGIN_GUID, configFile);
         Utils.Settings.Config.Load(); // just load this after setup all actions.