+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to BetterMagnets will be documented in this file.
 
+## `1.0.0`
+### Added
+- Only "good" magnets spawn by default (configurable with `OnlyGoodMagnets`).
+- Option to automatically add the Company Issued Magnet artifact at the start of each run (`CompanyIssuedMagnetEnabled`).
+- Magnets can be earned at the end of each run without losing experience.
+- Company Issued Magnet artifact now has a real good use.
+
 ## `0.0.1`
 ### Added
 - Early preview release.